Move normalization logic directly to FrameConsumerResultCallback
It's help to not ruin the API of *Consumer classes.
