+++ core/src/test/java/org/testcontainers/containers/output/FrameConsumerResultCallbackTest.java
@@ -1,0 +1,7 @@
+package org.testcontainers.containers.output;
+
+import static org.junit.Assert.*;
+
+public class FrameConsumerResultCallbackTest {
+
+}
